+++ src/index.ts
@@ -12,3 +12,4 @@ export * as Html from "./Html";
 export * as React from "./React";
 export * as Http from "./Http";
 export * as LocalStorage from "./LocalStorage";
+export * as Navigation from "./Navigation";